Rename go-aserto/directory to go-aserto/ds
