+++ ds/v2/directory.go
@@ -1,13 +1,12 @@
-package directory
+package ds
 
 import (
 	"github.com/aserto-dev/go-aserto"
 	"github.com/aserto-dev/go-aserto/internal/hosted"
-	des "github.com/aserto-dev/go-directory/aserto/directory/exporter/v3"
-	dis "github.com/aserto-dev/go-directory/aserto/directory/importer/v3"
-	dms "github.com/aserto-dev/go-directory/aserto/directory/model/v3"
-	drs "github.com/aserto-dev/go-directory/aserto/directory/reader/v3"
-	dws "github.com/aserto-dev/go-directory/aserto/directory/writer/v3"
+	des "github.com/aserto-dev/go-directory/aserto/directory/exporter/v2"
+	dis "github.com/aserto-dev/go-directory/aserto/directory/importer/v2"
+	drs "github.com/aserto-dev/go-directory/aserto/directory/reader/v2"
+	dws "github.com/aserto-dev/go-directory/aserto/directory/writer/v2"
 	"github.com/hashicorp/go-multierror"
 	"github.com/pkg/errors"
 	"google.golang.org/grpc"
@@ -26,9 +25,6 @@ type Client struct {
 
 	// Client for the directory exporter service.
 	Exporter des.ExporterClient
-
-	// Client for the directory model service.
-	Model dms.ModelClient
 
 	conns []*grpc.ClientConn
 }
@@ -54,7 +50,6 @@ func New(opts ...aserto.ConnectionOption) (*Client, error) {
 		Writer:   dws.NewWriterClient(conn),
 		Importer: dis.NewImporterClient(conn),
 		Exporter: des.NewExporterClient(conn),
-		Model:    dms.NewModelClient(conn),
 		conns:    []*grpc.ClientConn{conn},
 	}, nil
 }
