+++ ds/v2/export_test.go
@@ -1,7 +1,7 @@
-package directory
+package ds
 
 import (
-	"github.com/aserto-dev/go-aserto/directory/internal"
+	"github.com/aserto-dev/go-aserto/ds/internal"
 )
 
 func InternalConnect(conns *internal.Connections, cfg *Config) (*Client, error) {
